Adding Scripts for Event Handling, Managing Game Resources
Added Sound Manager, Event Manager, Game Manager, Player Manager. Still a work in progress
