--- Assets/Code/damage.cs
@@ -1,10 +1,0 @@
-// Digx7
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class damage : MonoBehaviour
-{
-    [Tooltip ("This amount will be added to the players current health.  Make it a (+) positive number for healing, and a (-) negative number for damage.")]
-    public int units = 0;
-}
