+++ Assets/Scripts/GameState.cs
@@ -1,0 +1,7 @@
+﻿// brought to you by Jenni
+public enum GameState {
+    None,
+    StartMenu,
+    Paused,
+    Playing,
+}
