+++ Assets/Scripts/GameState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d81c75b24a222842b313d233d528795
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
